pyproject.toml
[project]
name = "fact-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
    "google-cloud-bigquery>=3.29.0",
    "google-genai>=1.2.0",
    "streamlit>=1.42.1",
    "beautifulsoup4>=4.13.3",
    "requests>=2.32.3",
    "lxml>=5.3.1",
    "pdfkit>=1.0.0",
    "anthropic>=0.47.2",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
